third_party/blink/web_tests/platform/generic/css3/calc/table-calcs-expected.txt

Tests that CSS3 calc() can be used in tables

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS cell.offsetWidth is 100
PASS cell.offsetHeight is 100
PASS cell.offsetWidth is 100
PASS cell.offsetHeight is 100
PASS cell.offsetWidth is 100
PASS cell.offsetHeight is 100
PASS cell.offsetWidth is 100
PASS cell.offsetHeight is 100
PASS cell.offsetWidth is 100
PASS cell.offsetHeight is 100
PASS cell.offsetWidth is 100
PASS cell.offsetHeight is 100
PASS cell.offsetWidth is 100
PASS cell.offsetHeight is 100
PASS cell.offsetWidth is 100
PASS cell.offsetHeight is 100
PASS cell.offsetWidth is 100
PASS cell.offsetHeight is 100
PASS cell.offsetWidth is 100
PASS cell.offsetHeight is 100
PASS cell.offsetWidth is 100
PASS cell.offsetHeight is 100
PASS cell.offsetWidth is 100
PASS cell.offsetHeight is 100
PASS successfullyParsed is true

TEST COMPLETE
